Revert "Link with xs-openssl and load custom configuration"
This reverts commit 4263822d49aebc82fa25c53e8d35ba301cabe8bd.

Revert explanation:
When linked with xs-openssl instead of openssl, we haven't found a way
to make the XO sdn-controller work, always getting an "Unknown CA"
error. For now, we do not have a way to avoid weak ciphers in 8.2
because of this as we need a more recent openssl to support the
OPENSSL_CONF environment variable to force a cipher list.

For now, we move back to linking with the base openssl.

Signed-off-by: David Morel <david.morel@vates.tech>